Task force hotfix update from release

+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/externalreference/S_ExternalReference_StepDef.java
@@ -34,6 +34,8 @@ import de.metas.common.util.CoalesceUtil;
 import de.metas.cucumber.stepdefs.AD_User_StepDefData;
 import de.metas.cucumber.stepdefs.C_BPartner_Location_StepDefData;
 import de.metas.cucumber.stepdefs.C_BPartner_StepDefData;
+import de.metas.cucumber.stepdefs.DataTableRow;
+import de.metas.cucumber.stepdefs.DataTableRows;
 import de.metas.cucumber.stepdefs.DataTableUtil;
 import de.metas.cucumber.stepdefs.M_Product_StepDefData;
 import de.metas.cucumber.stepdefs.context.TestContext;
@@ -88,6 +90,7 @@ import static de.metas.externalreference.model.X_S_ExternalReference.TYPE_UserID
 import static org.adempiere.model.InterfaceWrapperHelper.newInstanceOutOfTrx;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.compiere.model.I_AD_User.COLUMNNAME_AD_User_ID;
+import static org.compiere.model.I_AD_User.COLUMNNAME_C_BPartner_ID;
 import static org.compiere.model.I_M_Shipper.COLUMNNAME_M_Shipper_ID;
 
 @RequiredArgsConstructor
@@ -186,6 +189,14 @@ public class S_ExternalReference_StepDef
 
 				externalReferenceRecord.setRecord_ID(shipper.getM_Shipper_ID());
 			}
+			else if (type.getCode().equals(BPartnerExternalReferenceType.BPARTNER.getCode()))
+			{
+				final String partnerIdentifier = DataTableUtil.extractStringOrNullForColumnName(row, COLUMNNAME_C_BPartner_ID);
+				assertThat(partnerIdentifier).isNotNull();
+
+				final I_C_BPartner partnerRecord = bpartnerTable.get(partnerIdentifier);
+				externalReferenceRecord.setRecord_ID(partnerRecord.getC_BPartner_ID());
+			}
 			else
 			{
 				throw new AdempiereException("Unknown X_S_ExternalReference.Type! type:" + typeCode);
@@ -263,11 +274,7 @@ public class S_ExternalReference_StepDef
 	@And("remove external reference if exists:")
 	public void remove_external_reference_if_exists(@NonNull final DataTable dataTable)
 	{
-		final List<Map<String, String>> tableRows = dataTable.asMaps(String.class, String.class);
-		for (final Map<String, String> tableRow : tableRows)
-		{
-			removeExternalReferenceIfExists(tableRow);
-		}
+		DataTableRows.of(dataTable).forEach(this::removeExternalReferenceIfExists);
 	}
 
 	@Given("metasfresh contains S_ExternalReferences")
@@ -377,13 +384,17 @@ public class S_ExternalReference_StepDef
 		externalReferenceRepository.save(externalReference);
 	}
 
-	private void removeExternalReferenceIfExists(@NonNull final Map<String, String> tableRow)
+	private void removeExternalReferenceIfExists(@NonNull DataTableRow row)
 	{
-		final String externalReference = DataTableUtil.extractStringForColumnName(tableRow, I_S_ExternalReference.COLUMNNAME_ExternalReference);
-		final ExternalSystemType externalSystemType = ExternalSystemType.ofValue(DataTableUtil.extractStringForColumnName(tableRow, "ExternalSystem"));
-		final ExternalSystem externalSystem = externalSystemRepository.getByType(externalSystemType);
-		final IExternalReferenceType referenceType = externalReferenceTypes.ofCodeNotNull(DataTableUtil.extractStringForColumnName(tableRow, I_S_ExternalReference.COLUMNNAME_Type));
+		final String externalReference = row.getAsString(I_S_ExternalReference.COLUMNNAME_ExternalReference);
+		final ExternalSystemType externalSystemType = ExternalSystemType.ofValue(row.getAsString("ExternalSystem"));
+		final ExternalSystem externalSystem = externalSystemRepository.getOptionalByType(externalSystemType).orElse(null);
+		if (externalSystem == null)
+		{
+			return;
+		}
 
+		final IExternalReferenceType referenceType = externalReferenceTypes.ofCodeNotNull(row.getAsString(I_S_ExternalReference.COLUMNNAME_Type));
 		externalReferenceRepository.deleteByReference(externalReference, externalSystem.getId(), referenceType);
 	}
